backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
asyncpg==0.29.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
mcp-sdk==0.1.0
pydantic-settings==2.1.0
openai==1.12.0